sw: changes for FP32 SSR SIMD implementation

+++ sw/applications/include/network.h
@@ -35,10 +35,10 @@ typedef struct network_t_ {
     uint32_t IN_CH2;
     uint32_t OUT_CH;
 
-    double *b;
-    double *W;
-    double *b_grad;
-    double *W_grad;
+    float *b;
+    float *W;
+    float *b_grad;
+    float *W_grad;
 
     double *images;
     uint32_t *targets;
